requirements.txt
streamlit>=1.30.0
groq>=1.0.0
google-generativeai>=0.3.0
google-genai>=1.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
requests>=2.31.0
yfinance>=0.2.36
python-dotenv>=1.0.0
pandas>=2.0.0
